Added a unit test that captures precision issues from arctan2, and the corresponding fix for is_inside_turbo. is_inside_naive, being naive, remains vulnerable to that issue

+++ is_inside_mesh.py
@@ -45,7 +45,7 @@ def is_inside_turbo(triangles, X):
 
 
 	# One generalized winding number per input vertex
-	ret = numpy.zeros(X.shape[0])
+	ret = numpy.zeros(X.shape[0], dtype = X.dtype)
 	
 	# Accumulate generalized winding number for each triangle
 	for U, V, W in triangles:
